refactor: move the code used for testing to lib.rs
